Implement the Mpris plugin
The Mpris plugin is for displaying metadata from media players that
implement the MPRIS D-Bus Interface Specification version 2.2.

The latest version of the specification can be found here:
http://specifications.freedesktop.org/mpris-spec/latest/index.html

When configured for a given player, (default vlc), the plugin should
display metadata in the statusline in the format "ARTIST - TITLE" for
the media player with the bus name `org.mpris.MediaPlayer2.[player]`.