i3focus bugfix: update section when con is focused
Responding only to the "focus" event ignores the new events that are
emitted on title changes.

Title change events are useful for cases such as knowing the title of
the tab you are on in Firefox, or knowing whether or not your vim
buffer has changes which is especially important when titlebars are
disabled.